Minor correction

+++ pb2csv.py
@@ -14,6 +14,8 @@ for entry in os.listdir(pbDataDir):
     if os.path.isfile(pbPath):
         csvName = os.path.splitext(os.path.basename(pbPath))[0] + ".csv"
         csvPath = os.path.join(csvDataDir, csvName)
+        if os.path.exists(csvPath):
+            continue
         myinput = open(pbPath, 'rb')
         myoutput = open(csvPath, 'wb')
         p = subprocess.Popen(gtfsToCsv, stdin=myinput, stdout=myoutput,
